introduced a pixels-per-line system (all references to sizes etc should be from 0 to 19 inclusive in the model now so that the view can then multiply these values by the pixels per line constant) and also refactored some methods in the model to use floats for greater accuracy

+++ Gizmoball Code/src/Model/Ball.java
@@ -9,13 +9,13 @@ public interface Ball {
 
     double getBallRadius();
 
-    double getXpos();
+    float getXpos();
 
-    void setXpos(double xpos);
+    void setXpos(float xpos);
 
-    double getYpos();
+    float getYpos();
 
-    void setYpos(double ypos);
+    void setYpos(float ypos);
 
     Circle getCircle();
